requirements.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.22
django-import-export==3.2.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==11.3.0
PyYAML==6.0
sqlparse==0.4.3
tablib==3.4.0
xlrd==2.0.1
xlwt==1.3.0
whitenoise==6.9.0
